# swagger-gateway-spec.yaml
openapi: 3.0.0
paths:
...
  /analyzer-gateway/verifier/resolve/{did}:
    get:
      description: このエンドポイントはDIDを解決します。入力としてDIDを受け取ります。
      operationId: ExternalApiController_hundleGetDidDocument
      parameters:
        - name: Accept
          required: false
          in: header
          description: 処理識別子
          examples:
            application/did+json:
              value: applicaiton/json
              description: DIDドキュメントのメディアタイプ（JSON）
            application/did+ld+json:
              value: applicaiton/did+ld+json
              description: DIDドキュメントのメディアタイプ（JSON-LD）。
            application/ld+json;profile="https://w3id.org/did-resolution":
              value: application/ld+json;profile="https://w3id.org/did-resolution"
              description: DID解決結果のメディアタイプ (JSON-LD).
          schema:
            type: string
        - name: did
          required: true
          in: path
          description: 解決するDIDを指定する
          schema:
            example: did:key:z6MknCCLeeHBUaHu4aHSVLDCYQW9gjVJ7a63FpMvtuVMy53T
            type: string
      responses:
        '200':
          content:
            application/did+json:
              schema:
                type: object
                description: The DID document (JSON representation).
                example:
                  id: did:indy:sovrin:WRfXPg8dantKVubE3HX8pw
                  verificationMethod:
                    - id: did:indy:sovrin:WRfXPg8dantKVubE3HX8pw#key-1
                      type: Ed25519VerificationKey2018
                      publicKeyBase58: H3C2AVvLMv6gmMNam3uVAjZpfkcJCwDwnZn6z3wXmqPV
            application/did+ld+json:
              schema:
                type: object
                description: The DID document (JSON-LD representation).
                example:
                  '@context': https://www.w3.org/ns/did/v1
                  id: did:indy:sovrin:WRfXPg8dantKVubE3HX8pw
                  verificationMethod:
                    - id: did:indy:sovrin:WRfXPg8dantKVubE3HX8pw#key-1
                      type: Ed25519VerificationKey2018
                      publicKeyBase58: H3C2AVvLMv6gmMNam3uVAjZpfkcJCwDwnZn6z3wXmqPV
            application/ld+json;profile="https://w3id.org/did-resolution":
              schema:
                description: The DID resolution result.
                type: object
                properties:
                  didDocument:
                    type: object
                  didResolutionMetadata:
                    type: object
                    additionalProperties: {}
                  didDocumentMetadata:
                    type: object
                    additionalProperties: {}
          description: DIDの解決に成功した
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: サポートされていない表現です。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: 正常に解決されましたが、DIDは無効化されています。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: DIDメソッドはサポートされていません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Resolver Serviceへリクエストを中継する
      tags: *ref_0
  /analyzer-gateway/verifier/status-lists/{listId}/status/{index}:
    get:
      operationId: ExternalApiController_hundleGetStatus
      parameters:
        - name: listId
          required: true
          in: path
          description: Status list lisの識別子
          schema:
            type: string
        - name: index
          required: true
          in: path
          description: Status listの状態が設定されている位置情報
          schema:
            type: number
      responses:
        '200':
          description: Status listの状態取得が正常に終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Status listの状態を取得する
      tags: *ref_0
  /analyzer-gateway/manager/status-lists:
    post:
      operationId: ExternalApiController_handleCreateStatus
      parameters:
        - name: X-Correlation-ID
          in: header
          description: 処理識別子
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusListCreateDto'
      responses:
        '201':
          description: Status listの登録が正常終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 認証は成功したが、この操作を実行する権限がない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: VCデータの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags: *ref_0
  /analyzer-gateway/manager/status-lists/{listId}/status/{index}:
    patch:
      operationId: ExternalApiController_handleUpdateStatus
      parameters:
        - name: listId
          required: true
          in: path
          description: Status list lisの識別子
          schema:
            type: string
        - name: index
          required: true
          in: path
          description: Status listの状態が設定されている位置情報
          schema:
            example: 20
            type: number
        - name: X-Correlation-ID
          in: header
          description: 処理識別子
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusListUpdateDto'
      responses:
        '200':
          description: Status listの状態更新が正常終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 認証は成功したが、この操作を実行する権限がない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags: *ref_0
  /analyzer-gateway/verifier/trusted-issuers/{did}:
    get:
      operationId: ExternalApiController_handleGetIssuer
      parameters:
        - name: did
          required: true
          in: path
          description: 発行者のDID
          schema:
            type: string
      responses:
        '200':
          description: Trusted listの状態取得が正常に終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustedListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Trusted listの状態を取得する
      tags: *ref_0
  /analyzer-gateway/admin/trusted-issuers:
    get:
      operationId: ExternalApiController_handleGetIssuers
      parameters:
        - name: X-Correlation-ID
          in: header
          description: 処理識別子
          required: false
          schema:
            type: string
      responses:
        '202':
          description: Trusted listの取得が正常終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustedListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 認証は成功したが、この操作を実行する権限がない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags: *ref_0
  /analyzer-gateway/manager/trusted-issuers:
    post:
      operationId: ExternalApiController_handleCreateIssuer
      parameters:
        - name: X-Correlation-ID
          in: header
          description: 処理識別子
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectDidRegistrationDto'
      responses:
        '201':
          description: Trusted listの登録が正常終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustedListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 認証は成功したが、この操作を実行する権限がない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags: *ref_0
  /analyzer-gateway/manager/trusted-issuers/{subjectDid}:
    patch:
      operationId: ExternalApiController_handleUpdateIssuer
      parameters:
        - name: subjectDid
          required: true
          in: path
          description: 処理依頼者のDIDを設定する
          schema:
            type: string
        - name: X-Correlation-ID
          in: header
          description: 処理識別子
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectDidUpdateDto'
      responses:
        '200':
          description: Trusted listの更新が正常終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustedListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 認証は成功したが、この操作を実行する権限がない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags: *ref_0
    delete:
      operationId: ExternalApiController_handleDeleteTrustedIssuer
      parameters:
        - name: subjectDid
          required: true
          in: path
          description: 処理依頼者のDIDを設定する
          schema:
            type: string
        - name: X-Correlation-ID
          in: header
          description: 処理識別子
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Trusted listの削除が正常終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustedListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 認証は成功したが、この操作を実行する権限がない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags: *ref_0
info:
  title: VC Analyzer System GateWay API
  description: API documentation for the VC Analyzer System management plane.
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
    ErrorServiceMetadata:
      type: object
      properties:
        serviceName:
          type: string
          description: サービス名
        version:
          type: string
          description: サービスバージョン
        timestamp:
          type: string
          description: 応答時間
        processingTimeMillis:
          type: number
          description: 処理時間
        correlationId:
          type: string
          description: 処理識別子
      required:
        - serviceName
        - version
        - timestamp
        - processingTimeMillis
        - correlationId
...
    ErrorResponse:
      type: object
      properties:
        serviceMetadata:
          description: サービスメタ情報
          allOf:
            - $ref: '#/components/schemas/ErrorServiceMetadata'
        error:
          description: エラー情報
          allOf:
            - $ref: '#/components/schemas/ErrorMainResponse'
      required:
        - serviceMetadata
        - error
...
    StatusListSuccessResponse:
      type: object
      properties:
        payload:
          description: ペイロード
          allOf:
            - $ref: '#/components/schemas/StatusListResponse'
        serviceMetadata:
          description: サービスメタ情報
          allOf:
            - $ref: '#/components/schemas/ServiceMetadataDto'
      required:
        - payload
    StatusListCreateDto:
      type: object
      properties:
        size:
          type: number
          example: '1000'
        statusPurpose:
          type: string
          example: revocation
          description: 省略時は、`revocation`
        bits:
          type: number
          example: 0
          description: >-
            `{specName}` が "bitstring" の場合で、statusPurposeがmessageの時のみ必須 (1, 2,
            3)。
        bitMapping:
          type: object
          example: 1
          description: '`ビット値の文字列とステータス文字列の対応マップ`'
      required:
        - size
        - bitMapping
    StatusListUpdateDto:
      type: object
      properties:
        status:
          type: string
          example: revoked
      required:
        - status
...
    TrustedListSuccessResponse:
      type: object
      properties:
        payload:
          description: ペイロード
          allOf:
            - $ref: '#/components/schemas/TrustedIssuerResponse'
        serviceMetadata:
          description: サービスメタ情報
          allOf:
            - $ref: '#/components/schemas/ServiceMetadataDto'
      required:
        - payload
        - serviceMetadata
    SubjectDidRegistrationDto:
      type: object
      properties:
        subjectDid:
          type: string
          description: 登録者のDID
          example: did:key:z6Mkiq1TQyWPezBTGoqNvcq9b6azWMyn6HwQmbjo5LLpai68
        policy:
          type: object
          description: 未サポート
      required:
        - subjectDid
    SubjectDidUpdateDto:
      type: object
      properties:
        validUntil:
          type: string
          description: ISO 8601フォーマット
          example: '2024-11-03T01:23:56.789Z'
        policy:
          type: object